move ChapterCollectibles to TokenList
